// legacy/v1/frontend/src/api/generated/models/dispatch-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 派遣ステータス
 * @export
 * @enum {string}
 */

export const DispatchStatus = {
    Preparing: 'preparing',
    Dispatched: 'dispatched',
    Returning: 'returning',
    Completed: 'completed',
    Recalled: 'recalled'
} as const;

export type DispatchStatus = typeof DispatchStatus[keyof typeof DispatchStatus];